+++ dwertheimer.Favorites/src/index.js
@@ -1,2 +1,26 @@
 // @flow
 export { setFavorite, openFavorite, removeFavorite } from './NPFavorites'
+
+export {
+  changePreset,
+  runPreset01,
+  runPreset02,
+  runPreset03,
+  runPreset04,
+  runPreset05,
+  runPreset06,
+  runPreset07,
+  runPreset08,
+  runPreset09,
+  runPreset10,
+  runPreset11,
+  runPreset12,
+  runPreset13,
+  runPreset14,
+  runPreset15,
+  runPreset16,
+  runPreset17,
+  runPreset18,
+  runPreset19,
+  runPreset20,
+} from './NPFavoritePresets'