+++ dwertheimer.Favorites/CHANGELOG.md
@@ -1,9 +1,13 @@
 # Favorites Plugin Changelog
 
+## [1.1.0] - 2023-08-25 @dwertheimer
+
+- Add capability to set a preset to run a favorite URL/X-Callback
+
 ## [1.0.1] - 2021-11-30 (@dwertheimer)
-### Added
-Minor tweak to the plugin.json to be more descriptive in Description 
+
+- Minor tweak to the plugin.json to be more descriptive in Description
 
 ## [1.0.0] - 2021-11-16 (@dwertheimer)
-### Added
+
 - Initial plugin functions: `/fave`, `/unfave`, `/faves`